refactor snd headers
